pyproject.toml
[project]
name = "slabvis"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "AcipenserSturio", email = "AcipenserSturio@users.noreply.github.com" },
]
dependencies = ["pandas>=2.2.3", "Pillow>=11.1.0"]
requires-python = ">=3.13"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

# [tool.pdm]
# distribution = true

[tool.pdm.scripts]
convert = "pdm run python ./src/convert.py"
draw = "pdm run python ./src/draw.py"
